Fuzzy Logic start
Started on the Fuzzy Logic. This is my interpretation of some of the things the assignment wants. It is not finished as some of the methods I added still need to be used and formatted to be printed. I could not import the library properly for the life of me so if anyone would like to figure that out before I do is more than welcome to.